+++ common/components/activity/activity-items-table.styles.ts
