+++ common/components/activity/activity-items-table.ts
@@ -8,15 +8,15 @@ import {
   property,
   PropertyValues
 } from 'lit-element';
-import {ActivityItemsTableStyles} from './acivity-items-table.styles';
-import {fireEvent} from '../../../../utils/fire-custom-event';
+import {ActivityItemsTableStyles} from './activity-items-table.styles';
+import {fireEvent} from '../../../utils/fire-custom-event';
 import {ActivityItemRow} from './activity-item-row';
 import './activity-item-row';
 import {AnyObject, InterventionActivityItem} from '@unicef-polymer/etools-types';
-import {callClickOnSpacePushListener} from '../../../../utils/common-methods';
+import {callClickOnSpacePushListener} from '../../../utils/common-methods';
 import {PaperTextareaElement} from '@polymer/paper-input/paper-textarea';
 import {translate} from 'lit-translate';
-import {sharedStyles} from '../../../../common/styles/shared-styles-lit';
+import {sharedStyles} from '../../styles/shared-styles-lit';
 
 @customElement('activity-items-table')
 export class ActivityItemsTable extends LitElement {
