+++ common/components/activity/get-total.helper.ts
