+++ intervention-results/results-structure/modals/activity-dialog/activity-data-dialog.ts
@@ -4,9 +4,9 @@ import '@unicef-polymer/etools-currency-amount-input';
 import '@polymer/paper-input/paper-textarea';
 import '@polymer/paper-toggle-button';
 import '@unicef-polymer/etools-dialog/etools-dialog.js';
-import './activity-items-table';
+import '../../../../common/components/activity/activity-items-table';
 import {gridLayoutStylesLit} from '../../../../common/styles/grid-layout-styles-lit';
-import {formatCurrency, getTotal} from '../../../../utils/get-total.helper';
+import {formatCurrency, getTotal} from '../../../../common/components/activity/get-total.helper';
 import {EtoolsRequestEndpoint, sendRequest} from '@unicef-polymer/etools-ajax/etools-ajax-request';
 import {getEndpoint} from '../../../../utils/endpoint-helper';
 import {interventionEndpoints} from '../../../../utils/intervention-endpoints';
@@ -14,7 +14,7 @@ import {getDifference} from '../../../../common/mixins/objects-diff';
 import {getStore} from '../../../../utils/redux-store-access';
 import './activity-timeframes';
 import {fireEvent} from '../../../../utils/fire-custom-event';
-import {ActivityItemsTable} from './activity-items-table';
+import {ActivityItemsTable} from '../../../../common/components/activity/activity-items-table';
 import {updateCurrentIntervention} from '../../../../common/actions/interventions';
 import {ActivityTimeFrames} from './activity-timeframes';
 import {formatServerErrorAsText} from '@unicef-polymer/etools-ajax/ajax-error-parser';
